pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "iotguard"
version = "1.0.0"
description = "Enterprise IoT command security platform"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "IoTGuard Team" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Internet :: Log Analysis",
    "Typing :: Typed",
]
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.32,<1",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",
    "redis>=5.2,<6",
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "python-jose[cryptography]>=3.3,<4",
    "passlib[bcrypt]>=1.7,<2",
    "structlog>=24.4,<26",
    "prometheus-client>=0.21,<1",
    "httpx>=0.28,<1",
    "google-generativeai>=0.8,<1",
    "paho-mqtt>=2.1,<3",
    "python-dotenv>=1.0,<2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8,<1",
    "mypy>=1.13,<2",
    "pre-commit>=4.0,<5",
    "bandit>=1.8,<2",
]
test = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.24,<1",
    "pytest-cov>=6.0,<7",
    "httpx>=0.28,<1",
    "faker>=33.0,<35",
    "aiosqlite>=0.20,<1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/iotguard"]

[tool.ruff]
target-version = "py311"
line-length = 99
src = ["src"]

[tool.ruff.lint]
select = [
    "E", "W",   # pycodestyle
    "F",         # pyflakes
    "I",         # isort
    "N",         # pep8-naming
    "UP",        # pyupgrade
    "S",         # bandit (via ruff)
    "B",         # flake8-bugbear
    "A",         # flake8-builtins
    "C4",        # flake8-comprehensions
    "RUF",       # ruff-specific
]
ignore = ["S101", "S105", "S106"]

[tool.ruff.lint.isort]
known-first-party = ["iotguard"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--strict-markers -ra --cov=iotguard --cov-report=term-missing"

[tool.bandit]
exclude_dirs = ["tests"]
